bots/lead-scraper/requirements.txt
googlemaps>=4.10
gspread>=6.0
google-auth>=2.0
requests>=2.31
beautifulsoup4>=4.12
python-dotenv>=1.0
